Move login components into login folder
