+++ frontend/components/login/OAuth/oauth_manager.js
